refactor: replace legacy cover page editor
